Move frontend components to own folder
